Some code clean up and changes to JMS messaging config.

+++ Manantiales/ManantialesModel/src/main/java/mx/ecosur/multigame/manantiales/util/AdjGraph.java
@@ -42,7 +42,7 @@ public class AdjGraph {
         Point ret = null;
         for (Point p : map.keySet()) {
             Integer i = map.get(p);
-            if (i.intValue() == node) {
+            if (i == node) {
                 ret = p;
                 break;
             }
